Change preview window reuse check to examine new buffer command
Previously, an existing preview window would be reused when its parent
buffer was replaced with a new buffer of the same file type. This change
reuses the preview window when the parent buffer is replaced if the new
buffer has a pipe preview command defined instead. If no command is
defined for the new buffer, the preview window is closed.